requirements.txt
Flask==2.0.2
yoyo-migrations==7.3.2
Flask-pymysql==0.2.3
# mysqlclient
# flask_mysqldb